requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
fastcluster==1.2.6
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.23.3
idna==3.7
ImageHash==4.3.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
PyWavelets==1.6.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
ruff==0.4.8
safetensors==0.4.3
scipy==1.13.1
sympy==1.12
tokenizers==0.19.1
torch==2.3.1+cpu
torchvision==0.18.1+cpu
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.9.0
urllib3==2.2.1
